+++ particula/activity/phase_separation.py
@@ -17,8 +17,6 @@ https://doi.org/10.5194/acp-19-13383-2019
 import numpy as np
 
 from particula.activity.machine_limit import safe_exp, safe_log
-from particula.activity.species_density import organic_density_estimate
-from particula.activity.binary_activity import activity_coefficients
 
 MIN_SPREAD_IN_AW = 10**-6
 Q_ALPHA_AT_1PHASE_AW = 0.99
@@ -241,62 +239,3 @@ def q_alpha(
             sigmoid_curve_parameter * (aw_series - a_w_sep + delta_a_w_sep)
         )
     )
-
-
-def biphasic_water_activity_point(
-    oxygen2carbon,
-    hydrogen2carbon,
-    molar_mass_ratio,
-    functional_group=None
-):
-    """
-    This function computes the biphasic to single phase
-    water activity (RH*100).
-
-    Args:
-    oxygen2carbon (np.array): An array representing oxygen2carbon values.
-    hydrogen2carbon (np.array): An array representing hydrogen2carbon values.
-    molar_mass_ratio (np.array): An array representing molar mass ratio values.
-    functional_group (str/list): The BAT functional group(s).
-
-    Returns:
-    np.array: The RH cross point array.
-    """
-
-    water_activity_cross_point = np.zeros_like(oxygen2carbon)
-
-    interpolate_step_numb = 200  # interpolation points
-    # mole_frac = np.linspace(1e-12, 1, interpolate_step_numb + 1)
-    mole_frac = np.logspace(-6, 0, interpolate_step_numb + 1)
-
-    for i, _ in enumerate(oxygen2carbon):
-        density = organic_density_estimate(
-            molar_mass_ratio[i],
-            oxygen2carbon[i],
-            hydrogen2carbon[i],
-            mass_ratio_convert=True)
-        activities = activity_coefficients(
-            molar_mass_ratio=molar_mass_ratio[i],
-            org_mole_fraction=mole_frac,
-            oxygen2carbon=oxygen2carbon[i],
-            density=density,
-            functional_group=functional_group
-        )
-
-        if np.isnan(activities[0]).any():
-            raise ValueError('water activity is NaN, check inputs')
-
-        phase_check = find_phase_separation(activities[0], activities[1])
-
-        if phase_check['phase_sep_check'] == 1:
-            water_activity_cross_point[i] = phase_check['upper_a_w_sep']
-        else:
-            water_activity_cross_point[i] = 0  # no phase separation
-
-    # Checks outputs with in physical limits
-    # round to zero
-    water_activity_cross_point[water_activity_cross_point < 0] = 0
-    # round max to 1
-    water_activity_cross_point[water_activity_cross_point > 1] = 1
-
-    return water_activity_cross_point
